requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
redis==4.5.5
kafka-python==2.0.2
click==8.1.3
pyyaml==6.0
pydantic==1.10.7
pandas==1.5.3
pyspark==3.3.1
transformers==4.30.2
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.2.2
joblib==1.2.0
tensorflow==2.12.0